test: consolidate and implement __main__.py tests
